+++ bindings/ruby/test/test_tree_tuner.rb
@@ -23,7 +23,7 @@ class CConfigSpaceTestTreeTuner < Minitest::Test
     tree = generate_tree(5, 0)
     ts = CCS::StaticTreeSpace.new(name: 'space', tree: tree)
     os = CCS::ObjectiveSpace.new(name: 'ospace')
-    v1 = CCS::NumericalParameter.new(lower: -Float::INFINITY, upper: Float::INFINITY)
+    v1 = CCS::NumericalParameter::Float.new(lower: -Float::INFINITY, upper: Float::INFINITY)
     os.add_parameter(v1)
     e1 = CCS::Variable.new(parameter: v1)
     os.add_objectives( {e1 => :CCS_OBJECTIVE_TYPE_MAXIMIZE} )
